# rust: speed up 'optimized-trie' program