After a question about why the trie program was slower via email, I
realized when I woke up the next morning that performance was likely
impacted by the speed of memory accesses themselves. And that if we
reduced the size of the table, we'd have a higher cache hit rate.
Indeed, in this patch, we do three things to shrink our in-memory size:

* Shrink the size of the 'counts' table by switching to u32. (A u16 is
too small to fit the largest count.)
* Shrink the size of the 'nodes' table by using a NODE_SIZE of 128. This
seems acceptable since we're allowed to assume the input is ASCII.
* Shrink the size of the 'nodes' table by using a u32 instead of a usize
(in practice, a u64) for our node IDs.

Each of these shrinks memory usage by a factor of 2. Overall, this
increases our cache hit rate and corresponds to a measurable
improvement. But still not quite as fast as the fastest hashmap
implementation:

    $ hyperfine './rust/optimized-trie/target/release/countwords-nodesize256 < kjvbible_x10.txt'
    Benchmark #1: ./rust/optimized-trie/target/release/countwords-nodesize256 < kjvbible_x10.txt
      Time (mean ± σ):     329.7 ms ±   3.5 ms    [User: 287.4 ms, System: 41.6 ms]
      Range (min … max):   322.8 ms … 333.5 ms    10 runs

    $ hyperfine './rust/optimized-trie/target/release/countwords-nodesize128-idusize < kjvbible_x10.txt'
    Benchmark #1: ./rust/optimized-trie/target/release/countwords-nodesize128-idusize < kjvbible_x10.txt
      Time (mean ± σ):     260.7 ms ±   2.5 ms    [User: 230.1 ms, System: 30.1 ms]
      Range (min … max):   258.0 ms … 266.9 ms    11 runs

    $ hyperfine './rust/optimized-trie/target/release/countwords-nodesize128-idu32 < kjvbible_x10.txt'
    Benchmark #1: ./rust/optimized-trie/target/release/countwords-nodesize128-idu32 < kjvbible_x10.txt
      Time (mean ± σ):     225.7 ms ±   2.9 ms    [User: 206.4 ms, System: 18.7 ms]
      Range (min … max):   221.6 ms … 230.3 ms    13 runs

    $ hyperfine './optimized-c < kjvbible_x10.txt'
    Benchmark #1: ./optimized-c < kjvbible_x10.txt
      Time (mean ± σ):     169.7 ms ±   8.2 ms    [User: 162.2 ms, System: 7.2 ms]
      Range (min … max):   156.4 ms … 178.0 ms    18 runs