@BurntSushi 2021-03-17
@wfxr can you show the output of `cat /proc/cpuinfo`? I wonder if you have a beefier CPU with a bigger cache. My CPU is from 2016 I believe.